deploy-aws/requirements.txt
acme==1.6.0
alembic==1.4.3
aniso8601==8.0.0
appdirs==1.4.3
apturl==0.5.2
argcomplete==1.12.1
asn1crypto==0.24.0
astroid==2.4.2
awscli==1.18.145
beautifulsoup4==4.8.2
blinker==1.4
boto3==1.15.6
botocore==1.18.6
Brlapi==0.7.0
certbot==1.8.0
certbot-dns-route53==1.8.0
certifi==2019.11.28
cfn-flip==1.2.3
chardet==3.0.4
chrome-gnome-shell==0.0.0
click==7.1.2
colorama==0.4.3
command-not-found==0.3
ConfigArgParse==0.13.0
configobj==5.0.6
crcmod==1.7
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==1.16.0
docutils==0.15.2
durationpy==0.5
entrypoints==0.3
eyeD3==0.8.10
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
future==0.18.2
gpg===1.13.1-unknown
hjson==3.0.2
html5lib==1.0.1
httplib2==0.14.0
idna==2.8
img2pdf==0.3.3
importlib-metadata==1.5.0
isort==5.5.2
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
josepy==1.2.0
kappa==0.6.0
keyring==18.0.1
keyrings.alt==3.4.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
linecache2==1.0.0
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
mock==3.0.5
more-itertools==4.2.0
mysqlclient==2.0.1
netifaces==0.10.4
numpy==1.17.4
oauth==1.0.1
oauthlib==3.1.0
olefile==0.46
packaging==20.3
parsedatetime==2.4
pbr==5.4.5
pexpect==4.6.0
photocollage==1.4.3
Pillow==7.0.0
pip-tools==5.3.1
placebo==0.9.0
protobuf==3.6.1
psutil==5.5.1
pyasn1==0.4.8
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyICU==2.4.2
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.1
python-dateutil==2.6.1
python-debian===0.1.36ubuntu1
python-editor==1.0.4
python-magic==0.4.16
python-slugify==4.0.1
pytz==2020.1
pyudev==0.21.0
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-toolbelt==0.8.0
requests-unixsocket==0.2.0
rsa==4.5
s3transfer==0.3.3
SecretStorage==2.3.1
simplejson==3.16.0
six==1.15.0
soupsieve==1.9.5
SQLAlchemy==1.3.19
ssh-import-id==5.10
system-service==0.3
systemd-python==234
testresources==2.0.0
text-unidecode==1.3
toml==0.10.1
tqdm==4.50.0
traceback2==1.4.0
troposphere==2.6.2
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unittest2==1.1.0
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.0.17
wadllib==1.3.3
webencodings==0.5.1
Werkzeug==0.16.1
wrapt==1.12.1
wsgi-request-logger==0.4.6
xkit==0.0.0
zappa==0.51.0
zipp==1.0.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.0.0
zope.interface==4.7.1
